+++ pymysqlreplication/constants/BINLOG.py
@@ -35,6 +35,7 @@ GTID_LOG_EVENT = 0x21
 ANONYMOUS_GTID_LOG_EVENT = 0x22
 PREVIOUS_GTIDS_LOG_EVENT = 0x23
 XA_PREPARE_EVENT = 0x26
+PARTIAL_UPDATE_ROWS_EVENT = 0x27
 
 # INTVAR types
 INTVAR_INVALID_INT_EVENT = 0x00
